Polish -- DRY refactor of the two generators via a generator base class.
